// src/app/modules/cadastros/cliente/enums/cadastros-cliente-enums.model.ts
/**********************************************************************************************
Code generated with MKL Plug-in version: 20.1.1
Code generated at time stamp: 2019-08-25T08:11:43.968
Copyright: Kerubin - logokoch@gmail.com

WARNING: DO NOT CHANGE THIS CODE BECAUSE THE CHANGES WILL BE LOST IN THE NEXT CODE GENERATION.
***********************************************************************************************/

import { SelectItem } from 'primeng/api';

export interface TipoPessoa extends SelectItem {
	
  label?: string;
  value: any;
  
}

export interface UF extends SelectItem {
	
  label?: string;
  value: any;
  
}